refactor: separação de rotas
