+++ fast_todo/routes/auth.py
@@ -1,0 +1,38 @@
+from typing import Annotated
+
+from fastapi import APIRouter, Depends, HTTPException
+from fastapi.security import OAuth2PasswordRequestForm
+from sqlalchemy import select
+from sqlalchemy.orm import Session
+
+from fast_todo.database import get_session
+from fast_todo.models import User
+from fast_todo.schemas import Token
+from fast_todo.security import create_access_token, verify_password
+
+router = APIRouter(prefix='/auth', tags=['auth'])
+
+OAuth2Form = Annotated[OAuth2PasswordRequestForm, Depends()]
+Session = Annotated[Session, Depends(get_session)]
+
+
+@router.post('/token', response_model=Token)
+def login_for_access_token(
+    form_data: OAuth2Form,
+    session: Session,
+):
+    invalid_access = HTTPException(
+        status_code=400, detail='Email ou senha inválido'
+    )
+
+    user = session.scalar(select(User).where(User.email == form_data.username))
+
+    if not user:
+        raise invalid_access
+
+    if not verify_password(form_data.password, user.password):
+        raise invalid_access
+
+    access_token = create_access_token(data={'sub': user.email})
+
+    return {'access_token': access_token, 'token_type': 'bearer'}
