+++ fast_todo/routes/users.py
@@ -1,0 +1,82 @@
+from typing import Annotated
+
+from fastapi import APIRouter, Depends, HTTPException
+from sqlalchemy import select
+from sqlalchemy.orm import Session
+
+from fast_todo.database import get_session
+from fast_todo.models import User
+from fast_todo.schemas import Message, UserList, UserPublic, UserSchema
+from fast_todo.security import get_current_user, get_password_hash
+
+router = APIRouter(prefix='/users', tags=['users'])
+
+Session = Annotated[Session, Depends(get_session)]
+CurrentUser = Annotated[User, Depends(get_current_user)]
+
+
+@router.post('/', status_code=201, response_model=UserPublic)
+def create_user(user: UserSchema, session: Session):
+    db_user = session.scalar(
+        select(User).where(User.username == user.username)
+    )
+
+    if db_user:
+        raise HTTPException(
+            status_code=400, detail='Esse usuário já foi cadastrado'
+        )
+
+    hashed_password = get_password_hash(user.password)
+
+    db_user = User(
+        email=user.email,
+        username=user.username,
+        password=hashed_password,
+    )
+
+    session.add(db_user)
+    session.commit()
+    session.refresh(db_user)
+
+    return db_user
+
+
+@router.get('/', status_code=200, response_model=UserList)
+def get_users(session: Session, skip: int = 0, limit: int = 100):
+    users = session.scalars(select(User).offset(skip).limit(limit)).all()
+    return {'users': users}
+
+
+@router.put('/{user_id}', status_code=200, response_model=UserPublic)
+def update_user(
+    user_id: int,
+    user: UserSchema,
+    session: Session,
+    current_user: CurrentUser,
+):
+    if current_user.id != user_id:
+        raise HTTPException(status_code=400, detail='Permissões insuficientes')
+
+    current_user.username = user.username
+    current_user.password = user.password
+    current_user.email = user.email
+
+    session.commit()
+    session.refresh(current_user)
+
+    return current_user
+
+
+@router.delete('/{user_id}', response_model=Message)
+def delete_user(
+    user_id: int,
+    session: Session,
+    current_user: CurrentUser,
+):
+    if current_user.id != user_id:
+        raise HTTPException(status_code=400, detail='Permissões insuficientes')
+
+    session.delete(current_user)
+    session.commit()
+
+    return {'detail': 'Usuário deletado'}
